+++ test/ecto_fragment_extras/convert_named_params_test.exs
@@ -1,12 +1,12 @@
-defmodule ConvertToEctoFragmentTest do
+defmodule ConvertNamedParamsTest do
   use ExUnit.Case
-  doctest EctoNamedFragment
+  doctest EctoFragmentExtras
 
-  alias EctoNamedFragment.ConvertToEctoFragment
-  alias EctoNamedFragment.Exceptions.CompileError
+  alias EctoFragmentExtras.ConvertNamedParams
+  alias EctoFragmentExtras.Exceptions.CompileError
 
   test "builds a fragment query string and splits params from kw list" do
-    assert ConvertToEctoFragment.call(
+    assert ConvertNamedParams.call(
              quote do
                "foo(#{:a}, #{:b})"
              end,
@@ -17,7 +17,7 @@ defmodule ConvertToEctoFragmentTest do
   end
 
   test "allows repeated param names" do
-    assert ConvertToEctoFragment.call(
+    assert ConvertNamedParams.call(
              quote do
                "foo(#{:a}, #{:b}, #{:a})"
              end,
@@ -31,7 +31,7 @@ defmodule ConvertToEctoFragmentTest do
     assert_raise KeyError,
                  "key :b not found in: [a: 0]",
                  fn ->
-                   ConvertToEctoFragment.call(
+                   ConvertNamedParams.call(
                      quote do
                        "foo(#{:a}, #{:b})"
                      end,
@@ -44,7 +44,7 @@ defmodule ConvertToEctoFragmentTest do
     assert_raise CompileError,
                  "names in named_fragment(...) queries must be atoms, got: a",
                  fn ->
-                   ConvertToEctoFragment.call(
+                   ConvertNamedParams.call(
                      quote do
                        "foo(#{a})"
                      end,
@@ -57,7 +57,7 @@ defmodule ConvertToEctoFragmentTest do
     assert_raise CompileError,
                  "named_fragment(...) expect a keyword list as the last argument, got: [\"a\"]",
                  fn ->
-                   ConvertToEctoFragment.call(
+                   ConvertNamedParams.call(
                      quote do
                        "foo(#{:a})"
                      end,