+++ CHANGELOG.md
@@ -18,7 +18,16 @@ Types of changes
 
 put your changes here
 
-## [0.1.0] - 2023-07-24
+## [0.3.0] - 2023-07-??
+
+* renamed library from `ecto_named_fragment` to `ecto_fragment_extras` because it contains a little more than just the named_fragment() macro now
+* added inline fragments which allow inlining fragment params into the query string:
+   
+  ```elixir
+  inline_fragment("coalesce(#{users.name}, #{^default_name})")
+  ```
+
+## [0.2.0] - 2023-07-24
 
 ### Changed